Refactor threading
