+++ kernel/src/main.rs
@@ -71,9 +71,9 @@ use kernel_api::memory::{Page, PhysicalAddress, VirtualAddress};
 use core::{future, mem, ptr};
 use core::cmp::{max, min};
 use core::num::NonZero;
-use core::task::{Poll, Waker};
+use core::task::{Poll};
 use core::time::Duration;
-use crate::threading::tcb::ThreadControlBlock;
+use crate::threading::ThreadControlBlock;
 use handoff_protection::HandoffWrapper;
 use hal::exception::DebugTy;
 use kernel_api::memory::{Frame};
@@ -88,6 +88,7 @@ use crate::hal::paging2::TTable;
 use crate::memory::paging::ktable;
 use crate::memory::watermark_allocator::WatermarkAllocator;
 use crate::task::executor::Executor;
+use crate::threading::{Waker, WakeReason};
 
 mod sync;
 mod memory;
